pyproject.toml
[tool.poetry]
name = "tg-bot"
version = "0.2.3"
description = "Xian Network Telegram Bot"
authors = ["endogen"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-telegram-bot = {extras = ["job-queue"], version = "^22"}
requests = "^2.31.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
psutil = "^5.9.8"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
segno = "^1.6.0"
qrcode-artistic = "^3.0.2"
plotly = "^5.18.0"
pytrends = "^4.9.2"
kaleido = "0.2.1"
numpy = "^1.26.4"
xian-tech-py = {path = "../xian-py", develop = true}
xian-tech-accounts = {path = "../xian-contracting/packages/xian-accounts", develop = true}
xian-tech-runtime-types = {path = "../xian-contracting/packages/xian-runtime-types", develop = true}
websockets = "^12.0"
pickledb = "^0.9.2"
dex-trade-api = "^0.1.0"
aiohttp = "^3.11.16"
aiosqlite = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.7"
mypy = "^1.11.2"
ruff = "^0.6.3"

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning"
]

[tool.mypy]
python_version = "3.11"
files = ["config.py", "constants.py", "tests"]
follow_imports = "skip"
exclude = "^(plg/|web\\.py|plugin\\.py|main\\.py)$"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = []

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "ASYNC"]
ignore = ["ASYNC109", "ASYNC230", "E501", "E722", "N806"]

[tool.ruff.lint.per-file-ignores]
"plg/**/res/*.py" = ["E741", "F821"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
